Fix BSPSceneManager, was broken by a recent signature change last week but missed ANSI compliance - remove anonymous structs from Vector2/3/4, Quaternion, ColourValue. 'val' is gone, use ptr() for pointer access, [] for array access Warning cleaning - OgreMain now builds with no warnings on VC8 /W3 Small tidy ups

+++ trunk/OgreMain/include/OgreBillboardSet.h
@@ -219,7 +219,7 @@ namespace Ogre {
         unsigned short mNumVisibleBillboards;
 
         /// Internal method for increasing pool size
-        virtual void increasePool(unsigned int size);
+        virtual void increasePool(size_t size);
 
 
         //-----------------------------------------------------------------------
@@ -287,7 +287,7 @@ namespace Ogre {
         /// Flag indicating whether the HW buffers have been created.
         bool mBuffersCreated;
         /// The number of billboard in the pool.
-        unsigned int mPoolSize;
+        size_t mPoolSize;
         /// Is external billboard data in use?
         bool mExternalData;
 
@@ -417,7 +417,7 @@ namespace Ogre {
             @param
                 size The new size for the pool.
         */
-        virtual void setPoolSize(unsigned int size);
+        virtual void setPoolSize(size_t size);
 
         /** Returns the current size of the billboard pool.
             @returns
